Refactor : Move ApiConstants and LocalDataConstants Change auth method from get to post for auth requests.
