--- identity_data/src/commonMain/kotlin/net/thechance/mena/identity/data/datasource/ApiConstants.kt
@@ -1,6 +1,0 @@
-package net.thechance.mena.identity.data.datasource
-
-object ApiConstants {
-    const val LOGIN = "identity/login"
-    const val REFRESH = "identity/refresh"
-}
