+++ identity_data/src/commonMain/kotlin/net/thechance/mena/identity/data/datasource/AuthRemoteDataSourceImpl.kt
@@ -2,7 +2,7 @@ package net.thechance.mena.identity.data.datasource
 
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
-import io.ktor.client.request.get
+import io.ktor.client.request.post
 import io.ktor.client.request.setBody
 import io.ktor.client.request.url
 import io.ktor.http.ContentType
@@ -27,11 +27,16 @@ private suspend inline fun <reified T, reified R> HttpClient.postJson(
     requestDto: T,
     path: String
 ): R {
-    return this.get {
+    return this.post {
         url(path)
         contentType(ContentType.Application.Json)
         setBody(requestDto)
     }.body()
 }
+object ApiConstants {
+    const val LOGIN = "identity/login"
+    const val REFRESH = "identity/refresh"
+}
+
