--- identity_data/src/commonMain/kotlin/net/thechance/mena/identity/data/datasource/LocalDataConstants.kt
@@ -1,6 +1,0 @@
-package net.thechance.mena.identity.data.datasource
-
-object LocalDataConstants {
-     const val ACCESS_TOKEN = "access_token"
-     const val REFRESH_TOKEN = "refresh_token"
-}
