+++ identity_data/src/commonMain/kotlin/net/thechance/mena/identity/data/datasource/LocalDataSourceImpl.kt
@@ -24,3 +24,7 @@ class LocalDataSourceImpl(
     }
 
 }
+object LocalDataConstants {
+    const val ACCESS_TOKEN = "access_token"
+    const val REFRESH_TOKEN = "refresh_token"
+}
